Change kernel return type to void (not return elapsed time any more)

+++ src/PyTorch_CUDA_extension/src/lstm.h
@@ -1,4 +1,4 @@
-float lstm_forward(
+int lstm_forward(
               THCudaTensor* h_data,
               THCudaTensor* x_data,
               THCudaTensor* c_data,
